Adds the upstream Solr "text_ja" field type
This adds the upstream Solr "text_ja" field type to be used for
Japanese language search.

Fixes an issue reported in
https://github.com/sul-dlss/sw_index_tests/issues/212
and tested in https://github.com/sul-dlss/sw_index_tests/blob/master/spec/cjk/japanese_series_spec.rb#L7-L10

Use case is a query of ユニークな should match 宣伝会議ユニークブックス.
But without this, な is not removed as a stoptag and the query is
unsuccessful.