Add configuration for Aardvark Earthworks cores
This adds solr configuration for a new Earthworks core using
the Aardvark schema, compatible with GeoBlacklight v4.

Once the core is created, we can index records into it using
a new traject config and proceed to update Earthworks to use
Geoblacklight v4.x. Later, the old core can be retired.